Add an opt in invalidate_active_sessions!  method to session timeout (#110)
* Add an opt in invalidate_activate_sessions!  method to session timeout

Spike on adding a configurable `invalid_active_sessions! method to the session_timeouts module. Configurable, but expects an `invalidate_sessions_before` timestamp column on user. Does not invalidate the session if `invalidate_sessions_before` is not set meaning that if it is deployed it won't log out currently logged in users.

* Add a test that you can login after calling `invalidate_active_sessions!`

There was concern that because `invalidate_sessions_before` is never cleared, there may be an issue where sessions would be cleared immediately after login. Add a spec to verify that sessions logged in after the `invalidate_sessions_before` timestamp can login and make additional requests without being cleared.

* Update the RREADME to document `invalidate_acitve_sessions!` method

* Move `invalidate_active_sessions!` method to no longer be protected

This method never should have been private as it was meant to be part of the public API if enabled

* Make the time check to invalidate the session more defensive

This would throw an error if session login_time and last_action_time were not set, be more defensive and default to current time if those values aren't set.