Use an error message from ruby-jwt exception messages
There is a suggestion that a message in the case not authenticated could be changeable
in the discussion: https://github.com/Sorcery/sorcery/pull/70#discussion_r158690311

But I decide to make use of a message sent from `JWT::DecodeError` because
it's difficult to make the message flexible because of the design.

Changes:

- `#jwt_require_auth` only rescues `JWT::DecodeError` because JWT::VerificationError is a subclass.
- `#jwt_decode` raises `JWT::DecodeError` or the error of the subclasses, and
does not rescue in it.
- `#jwt_not_authenticated` receives an argument

`JWT::DecodeError` are from here: https://github.com/jwt/ruby-jwt/blob/master/lib/jwt/error.rb