Check if mail object responds to delivery method
Instead of checking for an inheritance of ActionMailer::Base,
just see if the conffigure mailer and method respond to the configured
delivery method.

Since both the mailer and the willingniess to automatically send out
messages can be configured, the message should be send if it is
sendable, regardless of which Class the configured mailer inherits
from.