Support overwriting config per controller
It makes the configuration thread-safe by instantiating it.
This means that you can override the sorcery_config on the controller.

For example, you can use Admin instead of User on some controllers.

```ruby
class Admins::BaseController < ApplicationController
  before_action :set_admin_to_sorcery_config

  private

  def set_admin_to_sorcery_config
    sorcery_config.user_class = Admin
  end
end
```

This sample code is not enough because the current session key is a fixed value of `:user_id`.
I would like to address this problem in a next commit.