Removed the link to inch-ci(.org) badge in README.md because it wad dead
The link to the inch-ci badge in README.md because it was dead.

- Link: http://inch-ci.org/github/Sorcery/sorcery.svg?branch=master

It seems that `inch-ci.org` is not working,
and cannot access the site from the following URL.

- Main site: https://inch-ci.org/

The badge was not showing up in README.md for these reasons,
so I removed it to improve the documentation.