Fix rake not including the rubocop task
Honestly, I don't know if I want to even bother with rubocop in the legacy
codebase. The styling is being closely monitored in the rework, and code style
isn't really critical when it comes to security patches and bug fixes.

All to say, I may just disable the rubocop job for the legacy code again at some
point, if it gets annoying.